add default starting resources

+++ src/mechanics/Civilization.cpp
@@ -93,9 +93,9 @@ std::vector<const genie::Unit *> Civilization::swappableUnits(const uint16_t tas
     return m_taskSwapUnits[taskSwapGroup];
 }
 
-const std::unordered_map<genie::ResourceType, int> Civilization::startingResources() const
+const std::unordered_map<genie::ResourceType, float> Civilization::startingResources() const
 {
-    std::unordered_map<genie::ResourceType, int> ret;
+    std::unordered_map<genie::ResourceType, float> ret;
     for (size_t i=0; i<m_data.Resources.size(); i++) {
         if (i >= int(genie::ResourceType::NumberOfTypes)) {
             WARN << "Too many resources" << i;
